feat(showcase): one-click session export from the product + direct replay viewer
Export core moves from the CLI script into a backend-shared module
(apps/inno-agent/src/showcase/case-exporter.ts), so the product can
offer the same export the CLI has:

- POST /api/sessions/:id/showcase-export bundles the session (message
  stream + workspace files incl. bash-generated artifacts + wiki/profile
  keyframes) into <dataDir>/showcase-exports/cases, upserting its
  index.json. Title defaults to the session's custom/generated name.
- Session sidebar gains a clapperboard action ("导出为回放案例") that
  calls the endpoint and shows the cases dir (zh/en i18n).
- npm run showcase:view serves the built replay site with the exported
  cases dir overlaid: /cases/* resolves overlay-first with dist/cases
  as fallback and a merged index.json, auto-builds dist when missing,
  and opens the browser. No touching public/cases, no re-export into
  the site bundle.

The CLI keeps working unchanged (thin wrapper over the shared core).

Two exporter fixes surfaced by the endpoint flow:
- sanitizer: secret-prefix patterns (sk-/ghp_/Bearer …) now require a
  separator after the prefix — minified JS identifiers like
  skipHtmlTags or a.pkLongMethod() used to be rewritten to [REDACTED],
  breaking exported pages' inline scripts at runtime;
- sanitization now also maps the session's exact cwd to /workspace, so
  workspaces outside the standard container (e.g. .presets/*) produce
  message paths that match the replayed tree's relative paths.

Re-exported the published cases with the fixed sanitizer (false
positives removed; the one real api_key stays redacted). Verified
end-to-end: button endpoint -> overlay dir -> npm run showcase:view ->
full replay with generated HTML in the preview panel. 98/98 vitest.
